+++ typings/tutorial.d.ts
@@ -2,7 +2,7 @@ export type Maybe<T> = T | null
 
 export type TutorialConfig = {
   appVersions: TutorialAppVersions
-  testRunner: TutorialTestRunnerConfig
+  testRunner: TestRunnerConfig
   repo: TutorialRepo
   dependencies?: TutorialDependency[]
 }
@@ -52,10 +52,18 @@ export type StepActions = {
   subtasks?: string[]
 }
 
-export interface TutorialTestRunnerConfig {
+export interface TestRunnerArgs {
+  filter?: string
+  tap: string
+}
+
+export interface TestRunnerConfig {
   command: string
-  path?: string
-  actions?: StepActions
+  args?: TestRunnerArgs
+  path?: string // deprecated
+  directory?: string
+  actions?: StepActions // deprecated
+  setup?: StepActions
 }
 
 export interface TutorialRepo {